Correct webhost comments from nondeterminism to different output
Ignoring the issues with random option resolution on webhost, webhost
appears to be deterministic for a given seed, but it advances the
`random` module's RNG differently.

It may be possible to get the RNG of local generation and webhost
generation to match, but worlds should still not be using the `random`
module directly.